+++ libtitania-x3d/Titania/X3D/Components/Geospatial/GeoViewpoint.h
@@ -168,8 +168,8 @@ public:
 	getPosition () const;
 
 	virtual
-	Vector3f
-	getScreenScale (float, const Vector4i &) const final;
+	Vector3d
+	getScreenScale (double, const Vector4i &) const final;
 
 	///  @name Operations
